Added Filter for Listing Current / Pending Payments
In this commit, I also removed "total_costs" as a default attribute for ReportSheet and the index method for ReportSheets fires back a custom SQL with a few selected attributes. I'm not super happy with the solution, because for the PaymentOverview, I needed a completely different format of the ReportSheet than in the ReportSheetOverview and I wanted to cover both cases with one pretty index method, but the index method is now pretty messy.

Overall, the performance was so bad in my initial solution, this new approach had to be implemented, and it performs pretty good, so I'm fine with this solution.